requirements.txt
flask==2.0.1
flask-cors==3.0.10
bcrypt==3.2.0
python-dotenv==0.19.0
pyjwt==2.1.0
Werkzeug==2.0.1
